made 2d/ and 3d/ subdirectories for examples/
