+++ examples/2d/arrangement2d.jl
@@ -1,28 +1,10 @@
 using LinearAlgebraicRepresentation
+using Plasm,SparseArrays
 Lar = LinearAlgebraicRepresentation
-using Plasm
-import Base.show
-
-function show(filename)
-	V, EV = Plasm.svg2lar(filename)
-	#Plasm.view(Plasm.numbering(.075)((V,[[[k] for k=1:size(V,2)], EV])))
-	V, EV = Plasm.svg2lar(filename, flag=false)
-	#Plasm.view(Plasm.numbering(2)((V,[[[k] for k=1:size(V,2)], EV])))
-	return V, EV
-end
-
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/new.svg")
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/curved.svg")
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/twopaths.svg")
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/paths.svg")
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/boundarytest2.svg")
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/tile.svg")
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/interior.svg")
-#show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/holes.svg")
-
-V,EV = show("/Users/paoluzzi/Documents/dev/Plasm.jl/test/svg/holes.svg")
 
 
+# input generation
+V,EV = Lar.randomcuboids(30, .35)
 V = Plasm.normalize(V,flag=true)
 Plasm.view(Plasm.numbering(.05)((V,[[[k] for k=1:size(V,2)], EV])))
 
@@ -40,17 +22,9 @@ EW = Lar.cop2lar(copEV)
 W = convert(Lar.Points, V')
 hpcs = [ Plasm.lar2hpc(W,[EW[e] for e in comp]) for comp in bicon_comps ]
 Plasm.view([ Plasm.color(Plasm.colorkey[(k%12)==0 ? 12 : k%12])(hpcs[k]) for k=1:(length(hpcs)) ])
-Plasm.view(Plasm.numbering(.05)((W,[[[k] for k=1:size(W,2)], EW])))
-
-
-
 
 # computation of 2-cells and 2-boundary
 V, copEV, copFE = Lar.planar_arrangement_2(V, copEV, bicon_comps)
-
-
-
-
 
 # visualization of numbered arrangement
 Plasm.view( Plasm.numbering1(0.05)((V, copEV, copFE)) )
@@ -61,13 +35,14 @@ V = convert(Lar.Points, V')
 FVs = convert(Array{Lar.Cells}, triangulated_faces)
 Plasm.viewcolor(V::Lar.Points, FVs::Array{Lar.Cells})
 
-EVs = Lar.FV2EVs(copEV, copFE) # polygonal face boundaries
+# polygonal face boundaries
+EVs = Lar.FV2EVs(copEV, copFE) 
 EVs = convert(Array{Array{Array{Int64,1},1},1}, EVs)
 Plasm.viewcolor(V::Lar.Points, EVs::Array{Lar.Cells})
 
+# exploded polygons
 model = V,EVs
 Plasm.view(Plasm.lar_exploded(model)(1.2,1.2,1.2))
-
