+++ examples/2d/randomarrangement2d.jl
