+++ examples/2d/randomlines.jl
