+++ examples/3d/lario2obj.jl
