+++ examples/3d/simplexn_examples.jl
