+++ examples/3d/struct_examples.jl
