+++ examples/todebug/lar2obj2D.jl
