+++ examples/todebug/sweepline.jl
