+++ src/utilities.jl
@@ -809,6 +809,9 @@ Compute a *CDT* for each face of a `ChainComplex`. Return an `Array` of triangle
 function triangulate2D(V::Lar.Points, cc::Lar.ChainComplex)::Array{Any, 1}
     copEV, copFE = cc
     triangulated_faces = Array{Any, 1}(undef, copFE.m)
+    if size(V,2)==2 
+		V = [V zeros(size(V,1),1)]
+	end
 	
     for f in 1:copFE.m   
         edges_idxs = copFE[f, :].nzind
